vectorio: hoist the per-pixel invariants out of the fill loop
A Rectangle measured 1.2 us per pixel for four comparisons of shape
work: the loop did a mask lookup with division and modulus, the whole
screen-to-shape transform with its mirror and transpose branches, the
palette or converter lookup, and the depth switch for every pixel. The
transform is affine, so it is applied to two neighbouring pixels at the
start of each row and stepped by the difference, which keeps every
mirror and transpose case without repeating that logic. The mask bit is
walked, the colour is looked up only when the shape's index changes
(never, for the built-in shapes) and not at all when the shader dithers,
and the depth is decided once.

Cardputer, 64x64 shape, whole refresh including the SPI transfer:
Rectangle 5.04 ms to 3.36 ms, Polygon with 10 edges 9.98 ms to 8.85 ms,
against a 2.3 ms TileGrid floor. The polygon's remaining cost is its
per-pixel walk over the edges in get_pixel.

color_index was a uint16_t in every shape, so a 24-bit colour given to
a shape drawn through a ColorConverter lost its red byte, and the loop's
1-based-index correction was applied to the converter's colour too,
leaving it off by one in blue. The index is 32 bits, validated to
0..0xFFFFFF, and the converter gets the colour as given.

color_index was a uint16_t in every shape, so a 24-bit colour given to
a shape drawn through a ColorConverter lost its red byte: 0xFFFF00 came
out green, 0xFF0000 black. The index is 32 bits, validated to
0..0xFFFFFF, and the setters no longer take abs() of an unsigned value.

+++ shared-bindings/vectorio/Circle.h
@@ -12,7 +12,7 @@
 
 extern const mp_obj_type_t vectorio_circle_type;
 
-void common_hal_vectorio_circle_construct(vectorio_circle_t *self, uint16_t radius, uint16_t color_index);
+void common_hal_vectorio_circle_construct(vectorio_circle_t *self, uint16_t radius, uint32_t color_index);
 
 void common_hal_vectorio_circle_set_on_dirty(vectorio_circle_t *self, vectorio_event_t notification);
 
@@ -24,7 +24,7 @@ void common_hal_vectorio_circle_get_area(void *circle, displayio_area_t *out_are
 int16_t common_hal_vectorio_circle_get_radius(void *circle);
 void common_hal_vectorio_circle_set_radius(void *circle, int16_t radius);
 
-uint16_t common_hal_vectorio_circle_get_color_index(void *obj);
-void common_hal_vectorio_circle_set_color_index(void *obj, uint16_t color_index);
+uint32_t common_hal_vectorio_circle_get_color_index(void *obj);
+void common_hal_vectorio_circle_set_color_index(void *obj, uint32_t color_index);
 
 mp_obj_t common_hal_vectorio_circle_get_draw_protocol(void *circle);
