+++ shared-module/vectorio/Circle.h
@@ -15,7 +15,7 @@
 typedef struct {
     mp_obj_base_t base;
     uint16_t radius;
-    uint16_t color_index;
+    uint32_t color_index;
     vectorio_event_t on_dirty;
     mp_obj_t draw_protocol_instance;
 } vectorio_circle_t;
